[[Refactoring steps]]
Move update_time() inline in the code (used only once, it improves
readability of the source not to have to look what update_time
actually does).

Move the decision whether or not sendping() must ping the driver
to the place where it is invoked. When calling sendping(), it pings
the driver. Whether or not this is needed, is decided by the code
that calls sendping().

Fossil-ID: SVN r309