This changeset makes the upssched exec_cmd() error message more verbose.
It has been sitting in my tree for a while, mostly because I can't think of a
clean way to handle checking for the macros with autoconf. However, I checked
and this should work on most of the major Unix flavors (not sure about HP/UX,
but they seem to have good compatibility). Let me know if this causes more
problems than it solves.

Fossil-ID: SVN r437