[[Pre-release note]]
These are a fair number of changes late in the release cycle. Still I
think it is worth to include them in nut-2.0.5 since they fix a few
problems that are commonly experienced by users (bogus staleness
warnings). The corruption of the commandlist by the state_delcmd()
function also needs fixing. I it is too late now, feel free to reverse
them.

PS I did check all the rewritten state functions separately, to make
sure that in every conceivable situation they behave correctly and
that we're not introducing new bugs here.

PPS Yes, the use of recursion in many of these functions places more
demands on stackspace. I know. Still I think the better readability
(flatter) of the functions is worth it. We don't have enum lists with
thousands of parameters, so this should not be a big problem.

Fossil-ID: SVN r674