Include Subversion revision number in version number displays
>From Eric S. Raymond:

This is the patch to include SVN revision level in version number displays.
There should be two enclosures; the patch itself and a new source file.
common/upsversion.c.

upversion.c defines a single function. upsversion(), that returns a version
string for display.  Some Makefile trickery ensures that this file will
be recompiled whenever the project's SVN revision level has changed since
it was last built.

The other point of having a function is so this can encapsulate our policy
about how to compose a version string.  At the moment, the policy is this:
if the output of svnversion -n is empty, display only the configured 
version number.  If it's non-empty, display the output in parentheses 
after the configured version number.

Fossil-ID: SVN r920