[[Change in staleness detection]]
Update to reflect changes in the way we deal with dstate_dataok() and
dstate_datastale(). Backported from the trunk.

Historically it was needed to call dstate_dataok() regularly, to
prevent staleness warnings. This is no longer needed, now the server
will PING drivers it has not heard of recently. If they answer with
PONG, they won't be flagged stale anymore, no matter how long ago
dstate_dataok() was called.

Fossil-ID: SVN r997