Remove the stuff from alarm_commit() that prepends ALARM in 'ups.status'.
This proved to be rather counter productive and can be solved a lot
easier if we make sure that status_init() is not called between
alarm_init() and alarm_commit(). Will change the documentation later
this week, to reflect this change. Since so far very few drivers used
'ups.alarm', the impact is still limited.

Fossil-ID: SVN r1039