[[Improved matching bwhavior for usbhid-ups]]
* drivers/libhid.[ch],libshut.[ch],libusb.[ch]:
- Cut the libshut and libusb objects loose from libhid. If no HID
  functions are used (megatec_usb, tripplite_usb), you no longer need
  to include libhid anymore. The USB matcher functions (which were not
  really HID specific functions) are now within libusb.

* drivers/megatec_usb.c,tripplite_usb.c,Makefile.am:
- Applied the change above.

* drivers/usbhid.[ch]:
- The HIDOpenDevice() and HIDCloseDevice() functions no longer
  exist. These are replaced by direct calls to the communication
  subdrivers.
- The driver now gets a callback from libshut_open() and libusb_open()
  when a report descriptor has been retrieved from the USB/SHUT device
  and parses it and then makes a descision on whether this device is
  acceptable or not. If not, the next device that matches will be
  tried.

Apologies for the *really* big update, but the above required
reorganising many files. This patch now opens the way to much better
device matching for the usbhid-ups driver. The callback function in
the driver allows us to delay the decision whether or not the device
matches to *after* the subdriver specific ways to get to manufacturer,
model, and serial. I this is not enough to distinguish two identical
models from one another (by serial number), I doubt there is a way to
do this.

Fossil-ID: SVN r1076