drivers/usbhid-ups.[ch]: Changed the 'reconnect' option to 'matching'.
Behaviour is slightly different.

It wasn't really useful to only delay exact matching to after reading
the report descriptor (and use vendor specific formatting), not to
speak of the 'restart' option that would guarantee connection to a
different UPS than on startup. Instead, the first matching is now
always on the subdriver. After that, 'matching = normal (default)'
will behave as nut-2.2.0 (and before). When 'matching = delayed',
first the report descriptor will be read, vendor specific formatting
will be applied and then the regex(->exact) matchers (if available)
will be used.

Fossil-ID: SVN r1085