Add option 'matching = checksum'...
...that in addition to the 'normal' matching, will also compare the
SHA-1 checkum of the report descriptor with the one that was read on
startup.

Maybe calculating a SHA-1 checksum is a lot of overkill and instead we
could probably also store the complete report descriptor. On the other
hand, this probably won't happen too often, so the added computational
effort is likely to be insignificant.

Fossil-ID: SVN r1092