- Removed new "shutdown.restart" variable and changed back to 'shutdown.return'.
- Corrected shutdown function (should fallback to 'shutdown.reboot',
  not 'shutdown.return').

We should rework the variables and commands that deal with shutting
down the UPS. It certainly should be much clearer when a command can
be used with a command and when not. With the option of adding
parameters to commands, the differences between commands and variables
are not so clear-cut anymore. Until we have decided how to proceed
with this, it is probably best to make do with what we already have.

Fossil-ID: SVN r1114