Don't use the data we receive through input reports, ...
...but retreive it by polling for the feature report (only the report
number is used to determine which one to get).

This should fix the problems with faulty reports we receive over the
interrupt pipeline for some UPS'es, where the contents of the input
reports are broken, but the corresponding feature report is good (some
Belkin and Tripplite models, to name a few).

Fossil-ID: SVN r1208