[[Revert an incorrect change.]]
The changes in [[SVN:1210]] broke the calculation of the size of
reports, so this change is reverted.

Check if the HID Path ends in 0x00000000 and if so (we now see these
again), assume this variable isn't used and don't update it when we
receive an update over the interrupt pipeline.

Optimized the order in which we determine if a variable needs updating
so that the most restrictive tests are done first.

[[SVN:1210]] = 2008-01-02T20:34:26Z!adkorte-guest@alioth.debian.org

Fossil-ID: SVN r1221