Sneak preview for the new XML driver based on the 'neon' library.
Work in progress!

Compile with 'gcc -o netxml-ups -I/usr/include/neon -lneon
netxml-ups.c mge-xml.c' (change paths if needed). Run it with

	./netxml-ups http://mge-snmp.example.com:80 2>/dev/null

The portnumber is optional if it is equal to the default for the
mechanism (here http). When run, it will show a listing of the
parameters it has read from the UPS. The mapping to NUT values is not
yet implemented, it's just an example of how the 'neon' library can be
used for this purpose.

Fossil-ID: SVN r1313