If a HID path for a report ends in 0x00000000, it will not be made available.
These are continuations of previous objects (that we can't handle
anyway). Instead, we will only use them to calculate the size of the
report.

This basically is an improved version of [[SVN:1210]], that horribly
broke the calculation of the length of reports. Instead of ignoring
these reports completely, we now use them to calculate the length of
the report and only then discard them.

[[SVN:1210]] = 2008-01-02T20:34:26Z!adkorte-guest@alioth.debian.org

Fossil-ID: SVN r1336