Handle all communications with the server through upscli_sendline/readline...
...and use of select() to prevent blocking on I/O calls when the
server is not responding (while still allowing it a couple of seconds
to reply).

The above fixes a bug with upsmon ignoring DEADTIME when the
connection to the server is lost. Since we're using blocking I/O here,
a read() from the server would stall upsmon. Since we can't change to
non-blocking I/O (after writing a command to the server, the client
immediately reads a reply), the only option left is to use
select(). We do that now for all communication with the server.

Fossil-ID: SVN r1388