Performance improvements.
Since only write/read a single command (so a handful of bytes to/from
the server at a time) and we don't support pipelining commands, it is
not unreasonable to expect that this can be done in a single
write/read action. So instead of reading characters one by one, just
read whatever is available and let the parser deal with it. Increase
the timeout value to deal with slow servers, set the timeout for write
to zero (if we need to wait for buffer space to clear, the server is
probably already overloaded anyway).

Fossil-ID: SVN r1394