[[Support for overriding UPS variables.]]
Make it possible to permanently override a variable set by the UPS (if
it returns wrong static data) or to set a default value (if some
static data is missing, either in the driver or in the UPS).

This is a first shot at implementing this, for the time being as an undocumented
feature. Syntax is

	override.nut.variable.name = value
	default.nut.variable.name = value

The 'override.' means that the value will be set once and is marked
'immutable'. Any subsequent updates will be ignored. Use this if a UPS
(or driver) shows wrong data and you want to permanently set it to the
correct value. Obviously, this will only work for static data.

When a 'default.' value is set, the value is also set, but without
marking it 'immutable'.  This may be useful if some UPSes that are
supported by a driver set this value and some don't. The ones that
don't set this value (for instance 'battery.voltage.nominal') will
then use the default value, while the ones that so set it, will simply
overwrite the default.

Fossil-ID: SVN r1527