The return code of the ser_get_*() functions is now 0 on timeouts.
This allows drivers to tell the difference between an error that
occurred reading from the serial port and timeouts while reading from
the UPS. The ser_send() and ser_send_pace() functions return -1 on
errors, so the return value should be signed, not unsigned.

I will check all serial drivers over the next few days and will update
the use of these functions where necessary. I deliberately chose to
break them first, so that I can refer to this commit for the reason
for the change.

Fossil-ID: SVN r1542