Fix problem with reading replies on slow systems.
The ser_get_line() function is unreliable when reading multi-line
replies from the UPS. See docs/new-drivers.txt for an explanation
why. Basically this will only work if the driver is able to keep up
with the flow of data from the UPS and the select_read() function in
ser_get_line() reads one character at a time. In that case, switching
to ser_get_char() is more reliable, since that is not prone to losing
data after reading ENDCHAR.

Another problem is that the driver seems to expect that on
errors/timeout, the length of the returned string is 0 (strlen(buf) ==
0). This is not guaranteed to be the case (ser_get_char() will return
truncated replies on timeout for instance).

Last but not least if the length of the returned string is 0, this
doesn't mean that there is no more data left in the input buffer. So
flushing the input buffers was not guaranteed to work.

Fossil-ID: SVN r1549