Move USB device matching stuff into usb-common.
Although this is quite a big patch, it only moves things (doesn't add
nor change any functionality).

This is to prepare for the next release after nut-2.4.0, where usb_device_open()
and usb_device_close() will be available for all USB devices, so that we don't
have to duplicate this for every new USB driver.

Fossil-ID: SVN r1668