[[mge-hid fix]]
Comment out models that meet the criteria for automatic
generation of the model name and improve the handling
of concatenation of iProduct and iModel. Note that
we should NEVER remove entries, since later on we may
need to add device type specific overrides.

In any case, devices with a type other than
MGE_DEFAULT should NEVER be removed, since these are
already used in to match if an override or value
should be used or not.

Fossil-ID: SVN r1724