First shot of limiting the impact of [[SVN:2040]] to the mge-hid subdriver.
Use a separate function that can optionally be supplied for variables
where we need to have a both a 'forward' and 'reverse'
conversion. This is easier to maintain than a single function that
does both (and does away with (void *) returns that need to be casted
to the proper type).

[[SVN:2040]] = 2009-10-21T23:52:59Z!clepple+nut@gmail.com

Fossil-ID: SVN r2045