Remove configurable outlet and revert to classic polling mode...
...with the possibility to wake up through NSM subscription (fixed to
the main outlet, add 'subscribe' to ups.conf).

It would be nice to integrate NSM capabilities in a client or the
server (including listening to UDP broadcasts), but for a driver that
must be able to monitor three outlets at the same time and multiple
NMC's from the same host, this will not work.

Fossil-ID: SVN r2114