[[C portability fix.]]
Automatic variables must have a length that can be computed at compile
time rather than runtime. GCC will happily accept this C++
construction here, but this is not C.

Follow the same principle as used elsewhere in the code, by hard
coding this to PW_ANSWER_MAX_SIZE, the maximum size of replies we
expect.

Fossil-ID: SVN r2173