The upsConfigLowBattTime OID is in minutes
In order to convert this to seconds, we need to multiply by
60. Currently, the conversion factor is stored in the length field. So
we can either make this value ST_FLAG_RW (and not convert from minutes
to seconds) or show the correct value without being able to modify it.

If we set ST_FLAG_RW | ST_FLAG_STRING, the length field must show the
length of the string that needs to be passed by clients (and this
can't be done).

Fossil-ID: SVN r2738