Don't allocate storage for variables in snmp-ups.h
The header files should declare variables as extern, and they should be
allocated in only one .c file (not all of the files which include the .h
file). While this is not an immediate problem when only one .c file includes
the .h file, it still should be fixed in case that changes in the future.

Patch by Daniel Macks:
http://fink.cvs.sourceforge.net/viewvc/fink/experimental/dmacks/finkinfo/nut.patch?revision=1.1&view=markup

Fossil-ID: SVN r3449