Final fix to the tcsetattr() / tcgetattr() issue
The most likely cause of all spurious differences between what was set
on the port with tcsetattr() and what tcgetattr() shows are likely to do
with the c_local PENDIN flag, which is a status bit, not a control bit.
It will change when there's unread pending input, which can be quite
often on an APC UPS.

The right way to compare struct termios values is to clear the status
flags _after_ the tcsetattr() and of course after the tcgetattr() call
and then compare the result with what was set.

Also set NOKERNINFO, if available, as we don't want the UPS or noise on
the line to accidentally trigger status output back to the UPS, and
finally make sure IEXTEN is also cleared along with ISIG since it too
can cause weird things to happen.

This change also adds some debug code to show any differences in the
structures in a logical manner in debug output (and squashes one tiny
compiler warning) (patch from Greg A. Woods)

Fossil-ID: SVN r3603