bcmxcp: Add handling of BCMXCP_RETURN_ACCEPTED_PARAMETER_ADJUST and others in ACK block
Add support for handling more return statuses when exeucting commands, the most important
being BCMXCP_RETURN_ACCEPTED_PARAMETER_ADJUST, which means the command was executed.
The others added all handles cases where command was not executed, but you now get a
more detailed entry in log as to why it was not executed.