Provide retry options for upsdrvctl and driver(s)
As recently seen in Debian (bugs #694717 and #677143), it may be required to
have upsdrvctl retrying to start the driver in case of failure.  More
specifically, a mix of init system (V and systemd), udev and USB device(s) can
result in the /dev entry not being available at driver startup, thus resulting
in a general failure to start NUT.  This commit provides at least a way to
overcome this issue. A more suitable  solution will require more work on NUT
design.

This patch if based on Arnaud Quette proposal