Formalizing the end of the relationship with Eaton
The situation of the relationship with Eaton has evolved, and since 2011
Eaton does not support NUT anymore. This may still evolve in the future.
But for now, please do not consider anymore that buying Eaton products
will provide you with official support from Eaton, or a better level of
device support in NUT.