Add '-n no-fatal-on-missing-conf' option to upsmon
Do not fail (i.e., return EXIT_FAILURE 1) when no MONITOR line is
configured in upsmon.conf. This may be useful to not wrongly fail after
initial installation. This is the last of three commits related to
networkupstools/nut#156. However, a last patch will address the matching
modifications to the systemd service files