nutdrv_qx: give subdrivers a last chance to process the command
Add (and document) a new function ('preprocess_command()') to preprocess the command to be sent to the device, just before the actual sending and, in case of instant commands/setvars, after the 'preprocess()' function has been triggered (if appropriate).
As an example, this function can be useful to add to all commands (both queries and instant commands/setvars) a CRC or to fill the command of a query with some data.
Also, in qx_process(), address buf size vs item->answer size earlier.
Update all subdrivers accordingly, bump versions.