Fix dummy-ups for external value changes
dummy-ups allow to change the values of the publicated variables through
the standard upsrw tool. This method is handy to script value changes,
in a controlled way, compared to the dynamic version (using the TIMER
keyword in .dev files), which changes the values in a non controlled
way. Bump driver version to 0.14