Fix default SNMP retries and timeout
The previous patch was using the default values from Net-SNMP, which
are set to -1. When the user was not providing overriden values, this
was causing the driver to not be able to establish the communication
with the device. The default values are now fixed, as per documented
(i.e. 5 retries and timeout of 1 second). Also bump the driver version
to 0.74