Implement 3-phase alarm handling for SNMP
snmp-ups now allows to publish 3-phase alarms in ups.alarm, the same way
as with outlet.  Declaration of such alarms are done using "Lx.alarm".
info_lkp_t structures messages are shared templates with outlets, and
use the string formats to include the context (outlet or phase) and the
number (of the outlet or phase) in alarm messages.  These alarms are
then published in "ups.alarm", with the standard mechanism for alarm
publication