nutdrv_qx: improve protocol identification in 'voltronic-qs-hex'
Since 'V' protocol, in reality, never happens to use the encoded version of the reply to the QS query, but it always uses the plain version already implemented in 'voltronic-qs' subdriver, remove it from the identification process of 'voltronic-qs-hex' subdriver.
Also, remove some non-significant entries from the testing table and increase version number.