Extend namespace for Automatic Transfer Switch (ATS)
Extend NUT namespace to support a new type of power device: ATS - Automatic
Transfer Switch.  These devices are used to setup 2 power systems, such as
UPS, to power a single power supply system, and be able to automatically
transfer between the input sources in case of failure of the primary one.
The added variable are for now limited to 'input.source' and
'input.source.preferred', but may be extended if needed