Default to type NUMERIC for variables
Any variable that is not STRING, RANGE or ENUM is just a simple numeric value.
The protocol documentation (net-protocol.txt) was previously stating that "The
default <type>, when omitted, is integer." which was not fully true, since a
variable could also be a float.  Hence, the wording was changed to advertise
this, and that each driver is then responsible for handling values as either
integer or float.  Moreover, instead of returning a TYPE "UNKNOWN", return
"NUMERIC", which is more suitable, and aligned with the NUT protocol
specification