NUT QA: add Coverity static code analysis information
NUT is now a registered Coverity project, and can thus get static code analysis
from it. This is just a first step, and actions are needed to address the actual
defects and complete the project configuration.
Due thanks to Synopsys, Inc., mother firm of Coverity, for providing this
service to Opensource project