snmp-ups: update for daisychain
Update the daisychain implementation, as per the latest specification, including
considerations on "device.0". Alarms are also handled for the daisychain
devices. Initial code for settings and commands is also present, but will be
completed in subsequent commits, along with a final cleanup and potential
composite settings and commands on "device.0" (that are then applied to all
daisychain devices)