libusb 1.0 port: remove the check on USB device count
With the port to libusb 1.0, a sanity check was done on the available number of
USB devices to try. This however caused a behavior regression, since the only
device that NUT may have access to can be the only UPS connected. In this case,
the driver may exit upon trying to reconnect if the device is transiently not
available

Reference: https://github.com/networkupstools/nut/issues/300