gen-snmp-subdriver.sh: improvements to get data
The sysOID mechanism, used by this SNMP subdriver generator, only works if
we're pointed somehow in the right direction. For example, it doesn't work if
sysOID is '.1.3.6.1.4.1.705.1' and data is at '.1.3.6.1.4.1.534'. However, the
sysOID sometimes points too deep (.1.3.6.1.4.1.534.2.12, while data can be
obtained from .1.3.6.1.4.1.534). The SNMP subdriver generator now supports
ramping up until it can get data, even if it has to go up to the root (.1),
which can generate very verbose output. It is still better to get a lot than
nothing