dstate: when dealing with ranges, don't lose already-set flags
Ranges are now explicitly flagged as numbers, but the use of dstate_setflags() means that we substitute any previously-set flag with the ones we pass to that function, losing, for example, an already-set r/w status, with the result of making an affected var unchangeable.
Use instead the newly added dstate_addflags() function, which preserves already-set flags.