snmp-ups: Fix counter OID matching
In case of multiple similar sysOID (i.e. same sysOID in different subdrivers),
there is a counter matching OID (generally using the model OID). However, the
matching was causing a crash of snmp-ups in case where it was not the right MIB
mapping and the OID for model entry was not available

Signed-off-by: Arnaud Quette <ArnaudQuette@Eaton.com>