Initial integration of QA Regression Testsuite
This is the initial commit of the nut test script and the needed testlib,
originally developed for Ubuntu, and also used by Debian. This version was
modified to be usable on recent systems, that use systemd and can also adapt
to either nut-driver.service (current approach) and nut-driver.target.

Some work remains to make it usable to test a compiled version of NUT, since the
current approach relies on installing packages

Reference: https://github.com/networkupstools/nut/issues/3

Signed-off-by: Arnaud Quette <ArnaudQuette@Eaton.com>