Disable convert_deci() that was needed a decade ago
Left the old code in place, commented away. If devices with old NetXML pop up (those which served in XML 10x the measured value), fixes can be applied based on voltage and amps and power (order of magnitude comparison), or perhaps model names and/or firmware versions or release dates, if we get those.