Try to adapt paths in documentation
Documentation still widely uses the default paths source when compiled from
source. While it is feasible to adapt some at compilation time, it is not
desirable since, for example, the online documentation on networkupstools.org
has to be generic. So try to make these paths more generic when possible, and
otherwise make a note