configure.ac: Fix ${systemdsystemshutdowndir} detection
On non-Debian systems, ${libdir} is /usr/lib64, but systemd files are
installed below /usr/lib/systemd. Instead of assuming that the shutdown
directory is below ${libdir}, assume that it has the same parent
directory as ${systemdsystemunitdir}, which we already detect properly
using pkg-config. The proper fix would be a --with option for the
shutdown directory, but that would be an overkill.

Signed-off-by: Michal Marek <MichalMarek1@eaton.com>