Problem: nutscan_init() called before debug is enabled
Solution: Move the call to this stateless function to the point after we parse arguments and just before we'd start the actual scans. Later this can be extended to also consider which scans were requested and so which libs to load.

Thanks to @clepple for research in https://github.com/networkupstools/nut/issues/500

Signed-off-by: Jim Klimov <EvgenyKlimov@eaton.com>