libusb1: don't use the same var to iterate both devices and interface descriptor
From our old libusb.c, libusb1.c inherited a for loop that iterates over the interface descriptor in order to try and find the HID descriptor among its 'extra' bytes.
Unluckily, it uses the same var (i) that's now used, at the same time, for the iteration through the device list, in which it happens to be nested.
Since this obviously spells disaster, make these vars unique (and add a bit of debugging info).