Improve the SNMP subdriver generation script (DMF version) from upstream, Part 3
Provide a 3rd mode to allow an offline creation for a complete MIB walk.
In this mode, we get data from 1 file (numeric snmpwalk dump of the whole SNMP
tree). The sysOID is extracted from the dump, and only the pointed subtree is
used. A MIB file MUST be provided, and is used to produce the string SNMP walk.

Signed-off-by: Arnaud Quette <ArnaudQuette@Eaton.com>
Signed-off-by: Jim Klimov <JimKlimov@Eaton.com> (transplanting)