libusb: libusb_exit() when no device is found, not if libusb_init() fails
libusb_exit() should only be called after a successful call to libusb_init() (that is, if we don't want to segfault or the like), so don't call it when libusb_init() fails.
Also, always call libusb_exit() when no (acceptable) device is found, as the (indirect) call in upsdrv_cleanup() is not performed when we don't have a handle.