libusb: before attempting to reopen a device, close it even on Linux
Way back in 2006, the 'double free or corruption' bug was reported for USB drivers:
- https://alioth-lists.debian.net/pipermail/nut-upsdev/2006-January/000552.html
- https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=354305
- https://alioth-lists.debian.net/pipermail/nut-upsuser/2006-April/000833.html

Apparently, our call to usb_close() after a device disconnect was triggering the mischievous behaviour, and, as such, we ended up not calling usb_close() to avoid that.

It turns out the bug was actually not in how we closed devices, but in how we opened them (and, further, it never showed up on bcmxcp_usb, which didn't use the buggy code).

The bug was first introduced in 4f7dbf4a: global variables were removed from our libusb and drivers had then to call our libusb_open() function passing to it the address of the handle for the (already) opened device (which was previously one of the libusb globals that was being removed), so that it could store the value of the handle of the newly opened device in it... all good, if only we didn't forget to nullify the value stored in the caller's var when a specific device (e.g. the disconnected one) could not be found and (re)opened.
As such, after a device was considered disconnected:
- the first time the driver tried to reconnect and called usb_close(), the previously opened device was successfully closed,
- but then the driver called our libusb_open() function to reopen it and, while that device was still not available, if at least one other device was available and the last device in the list could be successfully usb_open()'d (allocating the memory for the handle and returning its address) and, as it was not the expected one, usb_close()'d (free()'ing the allocated memory), the caller was left with the address returned by usb_open() stored in the passed handle...
- and the next time it tried to reconnect, since it called usb_close() with that address (of an already free()'d memory), the disaster happened.
All of this actually happened with the help of our libhid, but mentioning all the passages would have complicated this (absolutely unnecessary and perfectly avoidable) explanation even more.

Then, in cc792e77, the workaround that's now being removed was introduced: it worked, but, targeting only the symptoms, it didn't cure the cause and made our libusb leak memory.
Finally, the bug was solved (albeit unawarely) in e335c6bc.

After more than 12 (twelve) years, it's now time to get rid of that workaround and to stop leaking memory on Linux, so call libusb_close() even there.

Also, in order to avoid possible future problems, move that call before library initialisation (and device discovery):
- we were already on the brink of experiencing problems with our old libusb-0.1 implementation, as we called usb_init(), usb_find_busses() and usb_find_devices() before usb_close()'ing the previously opened device, but usb_find_devices() free()'s no longer available devices and usb_close(), on some platforms, accesses members of that (now free()'d) device...
- so, now, we surely have to at least call libusb_close() before libusb_get_device_list() (luckily, we already do that), since our libusb-compat-1.0 calls usb_find_devices() there...
- and who knows... maybe a future buggy version of libusb-1.0 could do strange things even on libusb_init() with previously opened devices...
- and, to my mind, there isn't any good reason not to do it.

Finally, also reset the handle immediately after calling libusb_close(), since, before the reset at the end of the function takes place, there could be a call to exit(), which, in turn (by means of upsdrv_cleanup()), is likely to attempt to libusb_close() that same (already libusb_close()'d) handle, if it is not NULL.